Cleanup and renaming of traning optiosn tests and save and load test
